fix: fixed conflicts issues
